+++ src/ts/Views/PlayableMRAID.ts
@@ -12,6 +12,7 @@ import { Localization } from 'Utilities/Localization';
 import { Diagnostics } from 'Utilities/Diagnostics';
 import { IMRAIDViewHandler, MRAIDView } from 'Views/MRAIDView';
 import { IObserver0, IObserver1, IObserver2 } from 'Utilities/IObserver';
+import { SdkStats } from 'Utilities/SdkStats';
 
 export class PlayableMRAID extends MRAIDView<IMRAIDViewHandler> {
 
@@ -121,6 +122,8 @@ export class PlayableMRAID extends MRAIDView<IMRAIDViewHandler> {
         }
         this.createMRAID(container).then(mraid => {
             iframe.onload = () => this.onIframeLoaded();
+            SdkStats.setFrameSetStartTimestamp(this._placement.getId());
+            this._nativeBridge.Sdk.logDebug('Unity Ads placement ' + this._placement.getId() + ' set iframe.src started ' + SdkStats.getFrameSetStartTimestamp(this._placement.getId()));
             iframe.srcdoc = mraid;
             this._arFrameUpdatedObserver = this._nativeBridge.AR.onFrameUpdated.subscribe(parameters => this.handleAREvent('frameupdate', parameters));
             this._arPlanesAddedObserver = this._nativeBridge.AR.onPlanesAdded.subscribe(parameters => this.handleAREvent('planesadded', parameters));
@@ -197,6 +200,9 @@ export class PlayableMRAID extends MRAIDView<IMRAIDViewHandler> {
 
             this.showMRAIDAd();
         }
+
+        const frameLoadDuration = Date.now() - SdkStats.getFrameSetStartTimestamp(this._placement.getId());
+        this._nativeBridge.Sdk.logDebug('Unity Ads placement ' + this._placement.getId() + ' iframe load duration ' + frameLoadDuration + ' ms');
     }
 
     private showLoadingScreen() {